Implementation of new Camera3D type: Oblique
Original commit by qaptoR
I tried and failed many times to get the actual commit itself to play nicely with merging to version 4.2, but instead I'm just going to copy the changes over manually. Not the most clean solution, but it is the one that will work fastest and with most certainty.